backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary